requirements.txt
requests
click
loguru
fastapi
uvicorn[standard]
pydantic
python-dotenv
redis